requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
mlflow>=2.0.0
fastapi>=0.95.0
uvicorn>=0.20.0
pydantic>=1.10.0
pytest>=7.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
joblib>=1.2.0
pyyaml>=6.0
python-multipart>=0.0.5
requests>=2.28.0

